fix: split graphics manager module
